+++ Modules/Filtering/ImageIntensity/include/itkDiscreteGaussianDerivativeImageFunction.hxx
