+++ Modules/Filtering/ImageIntensity/include/itkDiscreteHessianGaussianImageFunction.hxx
