+++ Modules/Filtering/ImageIntensity/test/Baseline/itkDiscreteGaussianDerivativeImageFunctionTest01.png.cid
