+++ Modules/Filtering/ImageIntensity/test/Baseline/itkDiscreteGradientMagnitudeGaussianImageFunctionTest01.png.cid
