+++ Modules/Filtering/ImageIntensity/test/Baseline/itkDiscreteHessianGaussianImageFunctionTest.mha.cid
