+++ Modules/Filtering/ImageIntensity/test/CMakeLists.txt
@@ -2,6 +2,9 @@ itk_module_test()
 set(ITKImageIntensityTests
     itkBoundedReciprocalImageFilterTest.cxx
     itkCosImageFilterAndAdaptorTest.cxx
+    itkDiscreteHessianGaussianImageFunctionTest.cxx
+    itkDiscreteGradientMagnitudeGaussianImageFunctionTest.cxx
+    itkDiscreteGaussianDerivativeImageFunctionTest.cxx
     itkExpImageFilterAndAdaptorTest.cxx
     itkVectorRescaleIntensityImageFilterTest.cxx
     itkTernaryMagnitudeSquaredImageFilterTest.cxx
@@ -710,6 +713,101 @@ itk_add_test(
   COMMAND
   ITKImageIntensityTestDriver
   itkRoundImageFilterTest)
+itk_add_test(
+  NAME
+  itkDiscreteHessianGaussianImageFunctionTest
+  COMMAND
+  ITKImageIntensityTestDriver
+  --compare
+  DATA{Baseline/itkDiscreteHessianGaussianImageFunctionTest.mha}
+  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteHessianGaussianImageFunctionTest.mha
+  itkDiscreteHessianGaussianImageFunctionTest
+  DATA{${ITK_DATA_ROOT}/Input/cthead1.png}
+  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteHessianGaussianImageFunctionTest.mha
+  2.0
+  0.01
+  32)
+itk_add_test(
+  NAME
+  itkDiscreteGradientMagnitudeGaussianImageFunctionTest01
+  COMMAND
+  ITKImageIntensityTestDriver
+  --compare
+  DATA{Baseline/itkDiscreteGradientMagnitudeGaussianImageFunctionTest01.png}
+  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGradientMagnitudeGaussianImageFunctionTest01.png
+  itkDiscreteGradientMagnitudeGaussianImageFunctionTest
+  DATA{${ITK_DATA_ROOT}/Input/cthead1.png}
+  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGradientMagnitudeGaussianImageFunctionTest01.png
+  2.0
+  0.01
+  32
+  0
+  1)
+itk_add_test(
+  NAME
+  itkDiscreteGradientMagnitudeGaussianImageFunctionTest02
+  COMMAND
+  ITKImageIntensityTestDriver
+  --compare-MD5
+  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGradientMagnitudeGaussianImageFunctionTest02.png
+  de1904605e667018baaa3f73c5e5c220
+  itkDiscreteGradientMagnitudeGaussianImageFunctionTest
+  DATA{${ITK_DATA_ROOT}/Input/cthead1.png}
+  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGradientMagnitudeGaussianImageFunctionTest02.png
+  1.5
+  0.03
+  64
+  1
+  1)
+itk_add_test(
+  NAME
+  itkDiscreteGradientMagnitudeGaussianImageFunctionTestUseImageSpacingOff
+  COMMAND
+  ITKImageIntensityTestDriver
+  --compare-MD5
+  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGradientMagnitudeGaussianImageFunctionTestUseImageSpacingOff.png
+  de1904605e667018baaa3f73c5e5c220
+  itkDiscreteGradientMagnitudeGaussianImageFunctionTest
+  DATA{${ITK_DATA_ROOT}/Input/cthead1.png}
+  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGradientMagnitudeGaussianImageFunctionTestUseImageSpacingOff.png
+  1.5
+  0.03
+  64
+  1
+  0)
+itk_add_test(
+  NAME
+  itkDiscreteGaussianDerivativeImageFunctionTest01
+  COMMAND
+  ITKImageIntensityTestDriver
+  --compare
+  DATA{Baseline/itkDiscreteGaussianDerivativeImageFunctionTest01.png}
+  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGaussianDerivativeImageFunctionTest01.png
+  itkDiscreteGaussianDerivativeImageFunctionTest
+  DATA{${ITK_DATA_ROOT}/Input/cthead1.png}
+  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGaussianDerivativeImageFunctionTest01.png
+  1
+  2.0
+  0.01
+  32
+  0)
+itk_add_test(
+  NAME
+  itkDiscreteGaussianDerivativeImageFunctionTest02
+  COMMAND
+  ITKImageIntensityTestDriver
+  --compare-MD5
+  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGaussianDerivativeImageFunctionTest02.png
+  488a4bf5daa08cadb3e8b1018a3fe5c0
+  itkDiscreteGaussianDerivativeImageFunctionTest
+  DATA{${ITK_DATA_ROOT}/Input/cthead1.png}
+  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGaussianDerivativeImageFunctionTest02.png
+  2
+  1.5
+  0.03
+  16
+  1)
+
 
 set(ITKImageIntensityGTests itkBitwiseOpsFunctorsTest.cxx itkArithmeticOpsFunctorsTest.cxx)
