+++ Modules/Nonunit/Review/test/CMakeLists.txt
@@ -5,9 +5,6 @@ set(ITKReviewTests
     itkConformalFlatteningMeshFilterTest.cxx
     itkConformalFlatteningQuadEdgeMeshFilterTest.cxx
     itkDirectFourierReconstructionImageToImageFilterTest.cxx
-    itkDiscreteGaussianDerivativeImageFunctionTest.cxx
-    itkDiscreteGradientMagnitudeGaussianImageFunctionTest.cxx
-    itkDiscreteHessianGaussianImageFunctionTest.cxx
     itkFastApproximateRankImageFilterTest.cxx
     itkGridForwardWarpImageFilterTest.cxx
     itkImageFunctionTest.cxx
@@ -148,100 +145,6 @@ itk_add_test(
   64
   2
   0)
-itk_add_test(
-  NAME
-  itkDiscreteGaussianDerivativeImageFunctionTest01
-  COMMAND
-  ITKReviewTestDriver
-  --compare
-  DATA{${ITK_DATA_ROOT}/Baseline/Review/itkDiscreteGaussianDerivativeImageFunctionTest01.png}
-  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGaussianDerivativeImageFunctionTest01.png
-  itkDiscreteGaussianDerivativeImageFunctionTest
-  DATA{${ITK_DATA_ROOT}/Input/cthead1.png}
-  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGaussianDerivativeImageFunctionTest01.png
-  1
-  2.0
-  0.01
-  32
-  0)
-itk_add_test(
-  NAME
-  itkDiscreteGaussianDerivativeImageFunctionTest02
-  COMMAND
-  ITKReviewTestDriver
-  --compare-MD5
-  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGaussianDerivativeImageFunctionTest02.png
-  488a4bf5daa08cadb3e8b1018a3fe5c0
-  itkDiscreteGaussianDerivativeImageFunctionTest
-  DATA{${ITK_DATA_ROOT}/Input/cthead1.png}
-  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGaussianDerivativeImageFunctionTest02.png
-  2
-  1.5
-  0.03
-  16
-  1)
-itk_add_test(
-  NAME
-  itkDiscreteGradientMagnitudeGaussianImageFunctionTest01
-  COMMAND
-  ITKReviewTestDriver
-  --compare
-  DATA{${ITK_DATA_ROOT}/Baseline/Review/itkDiscreteGradientMagnitudeGaussianImageFunctionTest01.png}
-  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGradientMagnitudeGaussianImageFunctionTest01.png
-  itkDiscreteGradientMagnitudeGaussianImageFunctionTest
-  DATA{${ITK_DATA_ROOT}/Input/cthead1.png}
-  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGradientMagnitudeGaussianImageFunctionTest01.png
-  2.0
-  0.01
-  32
-  0
-  1)
-itk_add_test(
-  NAME
-  itkDiscreteGradientMagnitudeGaussianImageFunctionTest02
-  COMMAND
-  ITKReviewTestDriver
-  --compare-MD5
-  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGradientMagnitudeGaussianImageFunctionTest02.png
-  de1904605e667018baaa3f73c5e5c220
-  itkDiscreteGradientMagnitudeGaussianImageFunctionTest
-  DATA{${ITK_DATA_ROOT}/Input/cthead1.png}
-  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGradientMagnitudeGaussianImageFunctionTest02.png
-  1.5
-  0.03
-  64
-  1
-  1)
-itk_add_test(
-  NAME
-  itkDiscreteGradientMagnitudeGaussianImageFunctionTestUseImageSpacingOff
-  COMMAND
-  ITKReviewTestDriver
-  --compare-MD5
-  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGradientMagnitudeGaussianImageFunctionTestUseImageSpacingOff.png
-  de1904605e667018baaa3f73c5e5c220
-  itkDiscreteGradientMagnitudeGaussianImageFunctionTest
-  DATA{${ITK_DATA_ROOT}/Input/cthead1.png}
-  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteGradientMagnitudeGaussianImageFunctionTestUseImageSpacingOff.png
-  1.5
-  0.03
-  64
-  1
-  0)
-itk_add_test(
-  NAME
-  itkDiscreteHessianGaussianImageFunctionTest
-  COMMAND
-  ITKReviewTestDriver
-  --compare
-  DATA{${ITK_DATA_ROOT}/Baseline/Review/itkDiscreteHessianGaussianImageFunctionTest.mha}
-  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteHessianGaussianImageFunctionTest.mha
-  itkDiscreteHessianGaussianImageFunctionTest
-  DATA{${ITK_DATA_ROOT}/Input/cthead1.png}
-  ${ITK_TEST_OUTPUT_DIR}/itkDiscreteHessianGaussianImageFunctionTest.mha
-  2.0
-  0.01
-  32)
 itk_add_test(
   NAME
   itkFastApproximateRankImageFilterTest3
